refactor: mark test framework internal
