+++ examples/fabric/batch_job/test_fabric_batchjob_unit.py
@@ -1,5 +1,6 @@
 # flake8: noqa: E501
 import pytest
+from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
 from data_factory_testing_framework.models.pipeline import Pipeline
 from data_factory_testing_framework.state import (
     PipelineRunState,
@@ -7,7 +8,6 @@ from data_factory_testing_framework.state import (
     RunParameter,
     RunParameterType,
 )
-from data_factory_testing_framework.test_framework import TestFramework, TestFrameworkType
 
 
 @pytest.fixture
