+++ src/data_factory_testing_framework/__init__.py
@@ -1,5 +1,5 @@
 from data_factory_testing_framework._functions.functions_repository import FunctionsRepository
-from data_factory_testing_framework.test_framework import TestFramework, TestFrameworkType
+from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
 
 __all__ = [
     "TestFramework",
