+++ tests/functional/activity_outputs/test_set_activity_output.py
@@ -1,6 +1,6 @@
 import pytest
+from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
 from data_factory_testing_framework.state.dependency_condition import DependencyCondition
-from data_factory_testing_framework.test_framework import TestFramework, TestFrameworkType
 
 
 def test_execute_pipeline_activity_child_activities_executed(request: pytest.FixtureRequest) -> None:
