+++ tests/functional/append_variable_pipeline/test_append_variable_activity_pipeline.py
@@ -1,10 +1,10 @@
 from typing import List
 
 import pytest
+from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
 from data_factory_testing_framework.models.activities.append_variable_activity import AppendVariableActivity
 from data_factory_testing_framework.models.activities.set_variable_activity import SetVariableActivity
 from data_factory_testing_framework.state import RunParameter, RunParameterType
-from data_factory_testing_framework.test_framework import TestFramework, TestFrameworkType
 
 
 @pytest.mark.parametrize(
