+++ tests/functional/filter_activity_pipeline/test_filter_activity_pipeline.py
@@ -1,8 +1,8 @@
 import pytest
+from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
 from data_factory_testing_framework.models.activities.activity import Activity
 from data_factory_testing_framework.models.activities.filter_activity import FilterActivity
 from data_factory_testing_framework.state import RunParameter, RunParameterType
-from data_factory_testing_framework.test_framework import TestFramework, TestFrameworkType
 
 
 @pytest.mark.parametrize(
