+++ tests/unit/models/activities/control_activities/test_if_condition_activity.py
@@ -1,9 +1,9 @@
 import pytest
+from data_factory_testing_framework._test_framework import TestFramework, TestFrameworkType
 from data_factory_testing_framework.models.activities.if_condition_activity import IfConditionActivity
 from data_factory_testing_framework.models.activities.set_variable_activity import SetVariableActivity
 from data_factory_testing_framework.models.data_factory_element import DataFactoryElement
 from data_factory_testing_framework.state import PipelineRunState, PipelineRunVariable
-from data_factory_testing_framework.test_framework import TestFramework, TestFrameworkType
 
 
 def test_when_evaluated_should_evaluate_expression() -> None:
